Refactor capabilities
