+++ client-react/src/client/connectors/google_drive_v2/capabilities/index.js
@@ -1,0 +1,11 @@
+import createFolder from './create-folder';
+import deleteResource from './delete-resource';
+import download from './download';
+import rename from './rename';
+
+export default (apiOptions, { showDialog, hideDialog, forceUpdate }) => ([
+  createFolder(apiOptions, { showDialog, hideDialog, forceUpdate }),
+  deleteResource(apiOptions, { showDialog, hideDialog, forceUpdate }),
+  download(apiOptions, { showDialog, hideDialog, forceUpdate }),
+  rename(apiOptions, { showDialog, hideDialog, forceUpdate })
+]);
