--- client-react/src/client/connectors/google_drive_v2/context-menu-layout.js
@@ -1,3 +1,0 @@
-module.exports = {
-
-};
