+++ client-react/src/client/connectors/google_drive_v2/google-drive-utils.js
