+++ client-react/src/client/connectors/google_drive_v2/index.js
@@ -1,13 +1,11 @@
 import api from './api';
 import capabilities from './capabilities';
-import contextMenuLayout from './context-menu-layout';
 import listViewLayout from './list-view-layout';
 import viewLayoutOptions from './view-layout-options';
 
 export default {
   api,
   capabilities,
-  contextMenuLayout,
   listViewLayout,
   viewLayoutOptions
 };
